+++ actix-service/src/then.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use futures::{try_ready, Async, Future, Poll};
+use futures::{try_ready, Async, Future, IntoFuture, Poll};
 
 use super::{IntoNewService, NewService, Service};
 use crate::cell::Cell;
@@ -157,7 +157,10 @@ where
     type Future = ThenNewServiceFuture<A, B, C>;
 
     fn new_service(&self, cfg: &C) -> Self::Future {
-        ThenNewServiceFuture::new(self.a.new_service(cfg), self.b.new_service(cfg))
+        ThenNewServiceFuture::new(
+            self.a.new_service(cfg).into_future(),
+            self.b.new_service(cfg).into_future(),
+        )
     }
 }
 
@@ -185,8 +188,8 @@ where
         InitError = A::InitError,
     >,
 {
-    fut_b: B::Future,
-    fut_a: A::Future,
+    fut_b: <B::Future as IntoFuture>::Future,
+    fut_a: <A::Future as IntoFuture>::Future,
     a: Option<A::Service>,
     b: Option<B::Service>,
 }
@@ -201,7 +204,10 @@ where
         InitError = A::InitError,
     >,
 {
-    fn new(fut_a: A::Future, fut_b: B::Future) -> Self {
+    fn new(
+        fut_a: <A::Future as IntoFuture>::Future,
+        fut_b: <B::Future as IntoFuture>::Future,
+    ) -> Self {
         ThenNewServiceFuture {
             fut_a,
             fut_b,